services/structure-service/services/bootstrap/main.py
import os
import sys
import time


def check_env():
    print("Checking environment variables...")
    required_vars = ["DATABASE_URL"]
    missing = [v for v in required_vars if not os.getenv(v)]
    if missing:
        print(f"ERROR: Missing environment variables: {missing}")
        return False
    return True


def check_manifest():
    print("Checking ingestion manifest...")
    manifest_path = "ingest/manifest.jsonl"
    if not os.path.exists(manifest_path):
        print(f"WARNING: Manifest {manifest_path} not found. Creating empty.")
        os.makedirs(os.path.dirname(manifest_path), exist_ok=True)
        with open(manifest_path, "w") as f:
            pass
    return True


def check_db():
    print("Checking database connection...")
    # Real check would use sqlalchemy/redis client
    # For now, just check if we can import them
    try:
        import sqlalchemy
        import redis
    except ImportError:
        print("ERROR: Missing database drivers (sqlalchemy/redis)")
        return False
    time.sleep(1)  # Simulate connection time
    print("Database connected.")
    return True


def check_schemas():
    print("Verifying schemas...")
    # Simulate schema verification
    if not os.path.exists("schemas/ingest/document.json"):
        print("ERROR: document.json schema missing!")
        return False
    print("Schemas verified.")
    return True


def main():
    print("Starting system bootstrap...")

    if not check_env():
        sys.exit(1)

    if not check_db():
        sys.exit(1)

    if not check_schemas():
        sys.exit(1)

    if not check_manifest():
        sys.exit(1)

    print("Bootstrap complete. System ready.")


if __name__ == "__main__":
    main()
