services/structure-service/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "structure"
version = "0.1.0"
description = "R&D Orchestration Topology - Deterministic Kernels + LLM Interface"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }

dependencies = [
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "httpx>=0.27.0",
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-otlp>=1.20.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.2.0",
    "jsonschema>=4.21",
    "mypy>=1.8",
    "types-PyYAML",
    "types-requests",
    "types-jsonschema",
]

[tool.setuptools.packages.find]
include = [
    "gateway*",
    "router*",
    "validator*",
    "kernels*",
    "models*",
    "runtime*",
    "plugins*",
    "telemetry*",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "-o",
    "cache_dir=../../.cache/pytest/services-structure",
    "--cov=.",
    "--cov-report=term-missing",
    "--cov-report=xml",
    "--cov-fail-under=90",
]

[tool.coverage.run]
branch = true
source = ["."]
omit = [
    "tests/*",
    "models/ingest.py",
    "pipelines/*",
    "schemas/ingest/*",
    "kernels/experiment.py",
    "kernels/project.py",
    "kernels/data_summary.py",
    "kernels/constants.py",
    "kernels/units.py",
    "gateway/main.py",
    "gateway/compliance.py",
    "router/llm_extractor.py",
    "validator/loader.py",
    "validator/integrity.py",
    "runtime/orchestrator.py",
]

[tool.mypy]
python_version = "3.11"
cache_dir = "../../.cache/mypy/services-structure"
strict = true
show_error_codes = true
plugins = ["pydantic.mypy"]
disallow_untyped_decorators = false

exclude = [
    "^eval/",
    "^tests/",
    "^scripts/",
    "^services/bootstrap/",
    "^pipelines/",
    "^schemas/ingest/",
]

[[tool.mypy.overrides]]
module = [
    "httpx.*",
    "opentelemetry.*",
    "yaml.*",
    "fastapi.*",
    "starlette.*",
    "uvicorn.*",
]
ignore_missing_imports = true

# Legacy / generated kernels: typed incrementally
[[tool.mypy.overrides]]
module = "kernels.*"
ignore_errors = true

# Heuristic LLM helper — typed incrementally
[[tool.mypy.overrides]]
module = "router.llm_extractor"
ignore_errors = true

# Legacy routers / validators — tighten incrementally
[[tool.mypy.overrides]]
module = "router.gates"
ignore_errors = true

[[tool.mypy.overrides]]
module = "validator.normalizer"
ignore_errors = true

[[tool.mypy.overrides]]
module = "validator.loader"
ignore_errors = true

[[tool.mypy.overrides]]
module = "validator.integrity"
ignore_errors = true

[[tool.mypy.overrides]]
module = "validator.schema_validator"
ignore_errors = true

[[tool.mypy.overrides]]
module = "models.ingest"
ignore_errors = true

[[tool.mypy.overrides]]
module = "models.clarify"
ignore_errors = true

[[tool.mypy.overrides]]
module = "gateway.logging"
ignore_errors = true

[[tool.mypy.overrides]]
module = "router.classifier"
ignore_errors = true

[tool.pydantic-mypy]
init_forbid_extra = false
init_typed = true
warn_required_dynamic_aliases = true

[tool.ruff]
line-length = 100
cache-dir = "../../.cache/ruff"

[tool.ruff.lint]
# F401: unused imports (framework defines interfaces ahead of implementation)
# E402: module-level import not at top (tests need path manipulation)
# F841: unused variable (scaffolding for future implementation)
ignore = ["F401", "E402", "F841"]
select = ["E", "F", "W"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["E402"]
